// src/api/generated/schemas/getPermissions200DataItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */
import type { GetPermissions200DataItemPortal } from './getPermissions200DataItemPortal';
import type { GetPermissions200DataItemMethod } from './getPermissions200DataItemMethod';
import type { GetPermissions200DataItemPath } from './getPermissions200DataItemPath';
import type { GetPermissions200DataItemDescription } from './getPermissions200DataItemDescription';

export type GetPermissions200DataItem = {
  id: string;
  key: string;
  portal: GetPermissions200DataItemPortal;
  group: string;
  /** @nullable */
  method: GetPermissions200DataItemMethod;
  /** @nullable */
  path: GetPermissions200DataItemPath;
  /** @nullable */
  description: GetPermissions200DataItemDescription;
};